ml/fastapi/requirements.txt
fastapi
pydantic
joblib
pandas
numpy
pdfplumber
python-dotenv
fastapi-limiter
redis
uvicorn
gunicorn
google-generativeai
requests
tensorflow
scikit-learn
